moving new puppeteer provider to use the default file name
