+++ test/lib/app-provider.js
